Added better hit boxes to zombie, and toned down weapon and ammo crate hit boxes
Added individual box colliders to each part of the zombie to better detect if the player hits the zombie. Couldn't use mesh colliders like previously because they are most resource heavy and when spawning lots of zombies it could cause performance issues. Not sure it would actually affect the game on most computers, but better to optimize for performance any way. Went back to weapons and ammo crate to simply the mesh colliders into box colliders for each part, so still more accurate than before but now not so resource costly.