Zombies deal damage to player, and fixes
If the box collider for the zombie hand hits the player, the player takes damage. Changed attack range to closer match the reach of the zombie in the attack animation, added speed while attacking to slow down fast zombies when they attack so they don't push the player as much.